dropzone customization for single file input | fix the repeatation code of dropzone in braches-create,update

+++ app/Services/Impl/FileServiceImpl.php
@@ -70,7 +70,7 @@ class FileServiceImpl implements FileService
     }
 
     function deleteIfExists($filePath){
-        if (Storage::disk('public')->exists($filePath)) {
+        if ($filePath && Storage::disk('public')->exists($filePath)) {
             Storage::disk('public')->delete($filePath);
             return 1;
         }
